requirements.txt
fastapi[standard]
uvicorn[standard]
jinja2
pymongo
python-dotenv